requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
flower==2.0.1
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.3
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-mock==1.11.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
v==1
vine==5.1.0
wcwidth==0.2.9
yarl==1.9.2
